requirements.txt
asgiref==3.2.3
Django==3.0.1
django-jsonfield==1.4.0
djangorestframework==3.11.0
pkg-resources==0.0.0
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework-jwt
drf-yasg==1.20.0
Unipath
python-dotenv==0.15.0
django-crontab==0.7.1
django-allauth==0.44.0
django-cors-headers==3.5.0
django-filter==2.4.0
django-oauth-toolkit==1.3.3
django-extensions
django-schema-graph
djangorestframework-simplejwt==4.3.0